Updated #call in cli.rb
Called directly on the @@all array and removed first element using Ruby's .shift method because it's not a restaurant object. Created an array and called the self.get_name method from the scraper to fill it, then iterated over the @@all array and added each name to each restaurant. A little messy, but it works!